+++ lib/src/suggestions/suggestions_list.dart
@@ -1,10 +1,8 @@
 import 'dart:async';
 import 'dart:math';
+import 'package:drop_down_search_field/drop_down_search_field.dart';
 import 'package:drop_down_search_field/src/keyboard_suggestion_selection_notifier.dart';
 import 'package:drop_down_search_field/src/should_refresh_suggestion_focus_index_notifier.dart';
-import 'package:drop_down_search_field/src/suggestions/suggestions_box.dart';
-import 'package:drop_down_search_field/src/suggestions/suggestions_box_decoration.dart';
-import 'package:drop_down_search_field/src/type_def.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:pointer_interceptor/pointer_interceptor.dart';
@@ -17,6 +15,7 @@ class SuggestionsList<T> extends StatefulWidget {
   final TextEditingController? controller;
   final bool getImmediateSuggestions;
   final SuggestionSelectionCallback<T>? onSuggestionSelected;
+  final SuggestionMultiSelectionCallback<T>? onSuggestionMultiSelected;
   final SuggestionsCallback<T>? suggestionsCallback;
   final ItemBuilder<T>? itemBuilder;
   final IndexedWidgetBuilder? itemSeparatorBuilder;
@@ -45,6 +44,11 @@ class SuggestionsList<T> extends StatefulWidget {
   final KeyEventResult Function(FocusNode _, KeyEvent event) onKeyEvent;
   final bool hideKeyboardOnDrag;
   final bool displayAllSuggestionWhenTap;
+  final PaginatedSuggestionsCallback<T>? paginatedSuggestionsCallback;
+  final bool isMultiSelectDropdown;
+  final List<T>? initiallySelectedItems;
+  final SuggestionsBoxController? suggestionsBoxController;
+  final Widget? textFieldWidget;
 
   const SuggestionsList({
     super.key,
@@ -53,6 +57,7 @@ class SuggestionsList<T> extends StatefulWidget {
     this.intercepting = false,
     this.getImmediateSuggestions = false,
     this.onSuggestionSelected,
+    this.onSuggestionMultiSelected,
     this.suggestionsCallback,
     this.itemBuilder,
     this.itemSeparatorBuilder,
@@ -79,6 +84,11 @@ class SuggestionsList<T> extends StatefulWidget {
     required this.onKeyEvent,
     required this.hideKeyboardOnDrag,
     required this.displayAllSuggestionWhenTap,
+    this.paginatedSuggestionsCallback,
+    required this.isMultiSelectDropdown,
+    this.initiallySelectedItems,
+    required this.suggestionsBoxController,
+    this.textFieldWidget,
   });
 
   @override
@@ -93,6 +103,7 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
   late VoidCallback _controllerListener;
   Timer? _debounceTimer;
   bool? _isLoading, _isQueued;
+  bool _paginationLoading = false;
   Object? _error;
   AnimationController? _animationController;
   String? _lastTextValue;
@@ -100,6 +111,8 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
       widget.scrollController ?? ScrollController();
   List<FocusNode> _focusNodes = [];
   int _suggestionIndex = -1;
+  int pageNumber = 0;
+  final multiSelectSearchFieldFocus = FocusNode();
 
   _SuggestionsListState() {
     this._controllerListener = () {
@@ -198,6 +211,32 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
         _suggestionIndex = -1;
       }
     });
+
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (widget.paginatedSuggestionsCallback != null) {
+        _scrollController.addListener(() async {
+          if (_scrollController.position.pixels ==
+              _scrollController.position.maxScrollExtent) {
+            if (_isLoading ?? false) return;
+            _isLoading = true;
+            setState(() {
+              _paginationLoading = true;
+            });
+            final olderLength = this._suggestions?.length;
+            pageNumber += 1;
+            await invalidateSuggestions();
+            if (olderLength == this._suggestions?.length) {
+              pageNumber -= 1;
+            }
+            if (mounted) {
+              setState(() {
+                _paginationLoading = false;
+              });
+            }
+          }
+        });
+      }
+    });
   }
 
   Future<void> invalidateSuggestions() async {
@@ -221,7 +260,11 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
       Object? error;
 
       try {
-        suggestions = await widget.suggestionsCallback!(suggestion);
+        if (widget.paginatedSuggestionsCallback != null) {
+          suggestions = await widget.paginatedSuggestionsCallback!(suggestion);
+        } else {
+          suggestions = await widget.suggestionsCallback!(suggestion);
+        }
       } catch (e) {
         error = e;
       }
@@ -241,8 +284,8 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
           this._suggestions = suggestions;
           _focusNodes = List.generate(
             _suggestions?.length ?? 0,
-            (index) => FocusNode(onKeyEvent: (_, event) {
-              return widget.onKeyEvent(_, event);
+            (index) => FocusNode(onKeyEvent: (focusNode, event) {
+              return widget.onKeyEvent(focusNode, event);
             }),
           );
         });
@@ -266,7 +309,9 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
         (this._suggestions?.isEmpty ?? true) && widget.controller!.text == "";
     if ((this._suggestions == null || isEmpty) &&
         this._isLoading == false &&
-        this._error == null) return Container();
+        this._error == null) {
+      return Container();
+    }
 
     Widget child;
     if (this._isLoading!) {
@@ -289,6 +334,19 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
       }
     } else {
       child = createSuggestionsWidget();
+    }
+
+    if (widget.isMultiSelectDropdown) {
+      child = Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 10),
+            child: widget.textFieldWidget,
+          ),
+          Flexible(child: child),
+        ],
+      );
     }
 
     final animationChild = widget.transitionBuilder != null
@@ -392,47 +450,83 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
   }
 
   Widget defaultSuggestionsWidget() {
-    Widget child = ListView.separated(
-      padding: EdgeInsets.zero,
-      primary: false,
-      shrinkWrap: true,
-      keyboardDismissBehavior: widget.hideKeyboardOnDrag
-          ? ScrollViewKeyboardDismissBehavior.onDrag
-          : ScrollViewKeyboardDismissBehavior.manual,
-      controller: _scrollController,
-      reverse: widget.suggestionsBox!.direction == AxisDirection.down
-          ? false
-          : widget.suggestionsBox!.autoFlipListDirection,
-      itemCount: this._suggestions!.length,
-      itemBuilder: (BuildContext context, int index) {
-        final suggestion = this._suggestions!.elementAt(index);
-        final focusNode = _focusNodes[index];
-        return TextFieldTapRegion(
-          child: InkWell(
-            focusColor: Theme.of(context).hoverColor,
-            focusNode: focusNode,
-            child: widget.itemBuilder!(context, suggestion),
-            onTap: () {
-              // * we give the focus back to the text field
-              widget.giveTextFieldFocus();
+    Widget child = Stack(
+      children: [
+        ListView.separated(
+          padding: EdgeInsets.zero,
+          primary: false,
+          shrinkWrap: true,
+          keyboardDismissBehavior: widget.hideKeyboardOnDrag
+              ? ScrollViewKeyboardDismissBehavior.onDrag
+              : ScrollViewKeyboardDismissBehavior.manual,
+          controller: _scrollController,
+          reverse: widget.suggestionsBox!.direction == AxisDirection.down
+              ? false
+              : widget.suggestionsBox!.autoFlipListDirection,
+          itemCount: this._suggestions!.length,
+          itemBuilder: (BuildContext context, int index) {
+            final suggestion = this._suggestions!.elementAt(index);
+            final focusNode = _focusNodes[index];
+            return TextFieldTapRegion(
+              child: widget.isMultiSelectDropdown
+                  ? StatefulBuilder(
+                      builder: (context, setState) {
+                        final isSelected = widget.initiallySelectedItems
+                                ?.contains(suggestion) ??
+                            false;
+                        return CheckboxListTile(
+                          controlAffinity: ListTileControlAffinity.leading,
+                          title: widget.itemBuilder!(context, suggestion),
+                          value: isSelected,
+                          onChanged: (bool? checked) {
+                            // widget.controller?.text = widget.initiallySelectedItems
+                            //         ?.map((e) => e.toString())
+                            //         .join(', ') ??
+                            //     '';
+                            widget.onSuggestionMultiSelected!(
+                                suggestion, checked ?? false);
+                            setState(() {});
+                          },
+                        );
+                      },
+                    )
+                  : InkWell(
+                      focusColor: Theme.of(context).hoverColor,
+                      focusNode: focusNode,
+                      child: widget.itemBuilder!(context, suggestion),
+                      onTap: () {
+                        // * we give the focus back to the text field
+                        widget.giveTextFieldFocus();
 
-              widget.onSuggestionSelected!(suggestion);
-            },
+                        widget.onSuggestionSelected!(suggestion);
+                      },
+                    ),
+            );
+          },
+          separatorBuilder: (BuildContext context, int index) =>
+              widget.itemSeparatorBuilder?.call(context, index) ??
+              const SizedBox.shrink(),
+        ),
+        if (_paginationLoading)
+          const Align(
+            alignment: Alignment.bottomCenter,
+            child: CircularProgressIndicator(),
           ),
-        );
-      },
-      separatorBuilder: (BuildContext context, int index) =>
-          widget.itemSeparatorBuilder?.call(context, index) ??
-          const SizedBox.shrink(),
+      ],
     );
 
     if (widget.decoration!.hasScrollbar) {
       child = MediaQuery.removePadding(
         context: context,
         removeTop: true,
-        child: Scrollbar(
-          controller: _scrollController,
-          child: child,
+        child: Theme(
+          data: ThemeData(
+            scrollbarTheme: getScrollbarTheme(),
+          ),
+          child: Scrollbar(
+            controller: _scrollController,
+            child: child,
+          ),
         ),
       );
     }
@@ -449,35 +543,87 @@ class _SuggestionsListState<T> extends State<SuggestionsList<T>>
         final focusNode = _focusNodes[index];
 
         return TextFieldTapRegion(
-          child: InkWell(
-            focusColor: Theme.of(context).hoverColor,
-            focusNode: focusNode,
-            child: widget.itemBuilder!(context, suggestion),
-            onTap: () {
-              // * we give the focus back to the text field
-              widget.giveTextFieldFocus();
+          child: widget.isMultiSelectDropdown
+              ? StatefulBuilder(
+                  builder: (context, setState) {
+                    final isSelected = widget.controller?.text
+                            .contains(suggestion.toString()) ??
+                        false;
+                    return CheckboxListTile(
+                      title: widget.itemBuilder!(context, suggestion),
+                      value: isSelected,
+                      onChanged: (bool? checked) {
+                        // widget.controller?.text = widget.initiallySelectedItems
+                        //         ?.map((e) => e.toString())
+                        //         .join(', ') ??
+                        //     '';
+                        widget.onSuggestionMultiSelected!(
+                            suggestion, checked ?? false);
+                        setState(() {});
+                      },
+                    );
+                  },
+                )
+              : InkWell(
+                  focusColor: Theme.of(context).hoverColor,
+                  focusNode: focusNode,
+                  child: widget.itemBuilder!(context, suggestion),
+                  onTap: () {
+                    // * we give the focus back to the text field
+                    widget.giveTextFieldFocus();
 
-              widget.onSuggestionSelected!(suggestion);
-            },
-          ),
+                    widget.onSuggestionSelected!(suggestion);
+                  },
+                ),
         );
       }),
       _scrollController,
     );
 
     if (widget.decoration!.hasScrollbar) {
-      child = MediaQuery.removePadding(
-        context: context,
-        removeTop: true,
-        child: Scrollbar(
-          controller: _scrollController,
-          child: child,
+      child = Theme(
+        data: ThemeData(
+          scrollbarTheme: getScrollbarTheme(),
+        ),
+        child: MediaQuery.removePadding(
+          context: context,
+          removeTop: true,
+          child: Scrollbar(
+            controller: _scrollController,
+            child: child,
+          ),
         ),
       );
     }
 
+    child = Stack(
+      children: [
+        child,
+        if (_paginationLoading)
+          const Align(
+            alignment: Alignment.bottomCenter,
+            child: CircularProgressIndicator(),
+          ),
+      ],
+    );
+
     child = TextFieldTapRegion(child: child);
 
     return child;
+  }
+
+  ScrollbarThemeData? getScrollbarTheme() {
+    return const ScrollbarThemeData().copyWith(
+      thickness: WidgetStatePropertyAll(
+          widget.decoration?.scrollBarDecoration?.thickness),
+      thumbColor: WidgetStatePropertyAll(
+          widget.decoration?.scrollBarDecoration?.thumbColor),
+      radius: widget.decoration?.scrollBarDecoration?.radius,
+      thumbVisibility: WidgetStatePropertyAll(
+          widget.decoration?.scrollBarDecoration?.thumbVisibility),
+      crossAxisMargin: widget.decoration?.scrollBarDecoration?.crossAxisMargin,
+      mainAxisMargin: widget.decoration?.scrollBarDecoration?.mainAxisMargin,
+      interactive: widget.decoration?.scrollBarDecoration?.interactive,
+    );
   }
 }