+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## 1.1.2+2 - 12/02/2025
+- Hotfix release for Multi select dropdown.
+- Scroll Support for Selected Items with Mouse wheel.
+
+## 1.1.2+1 - 11/02/2025
+- Hotfix release for Multi select dropdown.
+
+## 1.1.2 - 11/02/2025
+- Added more support for Multi Select dropdown.
+- Configurations for Multi select dropdown.
+- Resolved issues requested by users.
+
+## 1.1.1 - 09/12/2024
+- Suggestion box Scrollbar customization.
+- Resolved issues requested by users.
+- Added More example videos.
+
+## 1.1.0 - 03/12/2024
+- Multiple item selection dropdown added.
+- Fixes for searchable dropdown.
+- Resolved issues requested by users.
+
+## 1.0.6 - 09/10/2024
+
+- Fixes for height and circular progress bar position for paginated list.
+- resolved `dart format .`.
+
+## 1.0.5 - 04/08/2024
+
+- Added support for Pagination in dropdown.
+- resolved `dart format .`.
+
 ## 1.0.4 - 08/01/2024
 
 - Support for closing suggestion box when tapped outside.